Move extensions on SubscriptionDescription to SubscriptionDescriptionExtensions. Do not pollute SubscriptionDescription with responsibilities that don't belong to it.
